Move rest of code from util.go and validator.go
