+++ pkg/artifactory/resource/configuration/configuration.go
@@ -2,8 +2,8 @@ package configuration
 
 import (
 	"github.com/go-resty/resty/v2"
-
 	"github.com/jfrog/terraform-provider-shared/util"
+	"gopkg.in/ldap.v2"
 )
 
 func SendConfigurationPatch(content []byte, m interface{}) error {
@@ -13,4 +13,20 @@ func SendConfigurationPatch(content []byte, m interface{}) error {
 		Patch("artifactory/api/system/configuration")
 
 	return err
+}
+
+func ValidateLdapDn(value interface{}, _ string) ([]string, []error) {
+	_, err := ldap.ParseDN(value.(string))
+	if err != nil {
+		return nil, []error{err}
+	}
+	return nil, nil
+}
+
+func ValidateLdapFilter(value interface{}, _ string) ([]string, []error) {
+	_, err := ldap.CompileFilter(value.(string))
+	if err != nil {
+		return nil, []error{err}
+	}
+	return nil, nil
 }
