+++ pkg/artifactory/resource/repository/remote/resource_artifactory_remote_nuget_repository.go
@@ -4,7 +4,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/repository"
-	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/utils"
 	"github.com/jfrog/terraform-provider-shared/util"
 )
 
@@ -61,7 +60,7 @@ func ResourceArtifactoryRemoteNugetRepository() *schema.Resource {
 	}
 
 	return repository.MkResourceSchema(nugetRemoteSchema, repository.DefaultPacker(nugetRemoteSchema), unpackNugetRemoteRepo, func() interface{} {
-		repoLayout, _ := utils.GetDefaultRepoLayoutRef("remote", packageType)()
+		repoLayout, _ := repository.GetDefaultRepoLayoutRef("remote", packageType)()
 		return &NugetRemoteRepo{
 			RemoteRepositoryBaseParams: RemoteRepositoryBaseParams{
 				Rclass:              "remote",
