+++ pkg/artifactory/resource/repository/remote/resource_artifactory_remote_vcs_repository.go
@@ -3,7 +3,6 @@ package remote
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/repository"
-	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/utils"
 	"github.com/jfrog/terraform-provider-shared/util"
 )
 
@@ -38,7 +37,7 @@ func ResourceArtifactoryRemoteVcsRepository() *schema.Resource {
 	}
 
 	return repository.MkResourceSchema(vcsRemoteSchema, repository.DefaultPacker(vcsRemoteSchema), UnpackVcsRemoteRepo, func() interface{} {
-		repoLayout, _ := utils.GetDefaultRepoLayoutRef("remote", packageType)()
+		repoLayout, _ := repository.GetDefaultRepoLayoutRef("remote", packageType)()
 		return &VcsRemoteRepo{
 			RemoteRepositoryBaseParams: RemoteRepositoryBaseParams{
 				Rclass:              "remote",
